+++ helpers/index.js
@@ -1,1 +1,3 @@
 export { default as HttpError } from "./HttpErrors.js";
+
+export { default as ctrlWrapper } from "./ctrlWrapper.js";
