+++ routes/api/contacts.js
@@ -1,100 +1,19 @@
 import express from "express";
+import ctrl from "../../controllers/contacts.js";
 
-import Joi from "joi";
-
-import {
-  listContacts,
-  getById,
-  removeContact,
-  addContact,
-  updateContact,
-} from "../../models/contacts.js";
-
-import { HttpError } from "../../helpers/index.js";
+import schema from "../../schemas/contacts.js";
+import validateBody from "../../middlewares/validateBody.js";
 
 const router = express.Router();
 
-const schema = Joi.object({
-  name: Joi.string()
-    .required()
-    .messages({ "any.required": "missing required name field" }),
-  email: Joi.string()
-    .required()
-    .messages({ "any.required": "missing required email field" }),
-  phone: Joi.string()
-    .required()
-    .messages({ "any.required": "missing required phone field" }),
-});
+router.get("/", ctrl.getAll);
 
-router.get("/", async (req, res, next) => {
-  try {
-    const allContacts = await listContacts();
-    res.json(allContacts);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", ctrl.getContactById);
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await getById(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", validateBody(schema), ctrl.add);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = schema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", ctrl.deleteById);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.status(200).json({ message: "contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    if (JSON.stringify(req.body) === "{}") {
-      throw HttpError(400, "missing fields");
-    }
-    const { error } = schema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { contactId } = req.params;
-
-    const result = await updateContact(contactId, req.body);
-
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:contactId", validateBody(schema), ctrl.updateById);
 
 export default router;
